Created the scenes for the game
Welcome - The first screen. User will input his/her name and choose either to create or join a game
Store - The user can buy other characters to play with
Lobby - If the user creates the game, s/he will be redirected to this screen and wait for the other player to join
Join - If the user wants to join a game, s/he will have to input the code of the game. After s/he will be redirected to the Lobby scene
Game - Both players will enter this scene to play the game
Winner - This will display the winner of the game and choose to either rematch or exit the game