UNITY: After the key is checked and is valid, the user is redirected to the Lobby scene, where the code (key) for the game and player 2's name are displayed.
FIREBASE: When the user clicks on Join Game and the key is valid, a new child named Player 2, with the name of player 2, is created under the same key of Player 1.

BUG #1: User must click twice on the Join Game in order to add player 2 in Firebase and redirect the user.