UNITY: If both players submit their results before the timer expires, the timer will stop and the results will be shown. If the timer expires, the results will be shown. If one or both players did not submit any choice, "None" will be set as their choice. Also, points are added to the player according to the results.
BUG #2: Till now, the results are shown as text. If these are filled with each player's choice and after the score is added will be set to empty, the score will be added by one while the choices will be removed immediatly from the screen. On the other hand, if these are filled and then they are not set to empty, the score will continue to increase while the text will remain on the screen.